Update errors returned
We return 422 instead of 400 when we can't create more voters for a
given election. If the signature is invalid, we're returning 200 with a
template that explains the signature is invalid to the voter opening the
page. We decided to do that since our API will be used by people on
browsers instead of being consumed by another system.